login: check for empty AWS accounts in response
An invalid SAML assertion sent to the AWS SAML signin page or any other
non-transport error (eg. 5xx status codes) would result in the prompter
in an infinite loop as it would be passed an empty list.

This just checks that we got at least 1 account back from AWS before
prompting the user.