Update exec to perform role assumption from profile.
The method created by @85matthew worked for a large majority of tools
but we found that some don't properly utilize the AWS_PROFILE env var.
In particular the Sops tool has issues.

This update makes it such that saml2aws behaves more like some other
tools in this space (aws-vault and aws-okta). When the `--exec-profile`
flag is specified saml2aws now performs the role assumption via the STS
client from the AWS SDK. This means it will now to multiple levels of
role assumption _before_ executing the given command. Because saml2aws
is performing the assumption it no longer export the AWS_PROFILE env
vars and the credentials that it export are fully fledged credentials
for the target profile.