Add support for new Google Login Page Formats
Fixes #203 and #444

Google has been changing a number of their login pages recently. As a
result of that, their CAPTCHA page doesn't behave exactly as before, nor
do some of the login pages themselves.

This change is designed to be backwards compatible with the existing
Google login and captcha pages, as it appears that the changes are not
completely propogated through their systems at this time.

This commit contains a fair number of changes, among which are:

- Support for generating absolute URLs from relative paths
    - Some of the fields being references in `src` and `action`
    attributes are now relative, so we need to expand them
- Support for logging into a final password page after the CAPTCHA is
complete, as this behaviour now seems different