Allow for govcloud saml redirect
Currently the SAML use the default commercial partition for its calls.
This has the side-effect of preventing govcloud from working.
I used the fastest way to go there.

But we could parse the XML to use the saml2p:Response->Destination to use the correct target without having to rely on a `grep` of the content

And the region/partition could also be infered in a similar way.