Fix use of OKTA as an MFA provider setting
Saml2aws compares the configured MFA factor (with --mfa or the MFA
configuration option) with the display name of the availabile factors.
For the Okta Verify TOTP factor, the display name is "Okta MFA
Authentication", but the allowed configuration option is "OKTA", and so
when you set --mfa=OKTA (or the equivalent in the configuration file),
it doesn't match, and you get a push notification from Okta Verify
instead.

This change just makes the prefix comparison case insensitive, so
setting `--mfa=OKTA` will match the "Okta MFA Authentication" option,
and you are prompted for a TOTP code from the Okta Verify app.