Change import path to include /v2. Fixes #417.
According to
https://github.com/golang/go/wiki/Modules#semantic-import-versioning, when you
bump a major version in a package beyond v0 or v1, imports need to be renamed
to include a major versions suffix. If someone tries to import a package that
does not follow this scheme, `go get` will error with a message like:
`version "v2.26.2" invalid: module contains a go.mod file, so major version
must be compatible: should be v0 or v1, not v2`.

This patch fixes the import path throughout the tool to include the `/v2`
suffix. This was done using https://github.com/marwan-at-work/mod, as well as
hand-tweaking import paths that were not corrected by the tool, due to being
platform-specific.