Adding option to write to different cred file.
The AWS SDK reads to the default AWS credential file, but does not
handle natively the expiry timestamp. The credential_process allows the
cred renewals when needed, but as the credential file takes precedence,
the credential provider never triggers.

This feature allows saml2aws to write to a different credential file,
which it will use as credential cache and not "conflict" with the AWS
SDK.

The change includes the command line option and the saml2aws
configuration file item.