Adding SAML Assertion caching feature
You can now cache the SAML assertion after authentication using the flag
`--cache-saml`. You can use the flag at `configure` time and it will be
set on subsequent `login` and `list-roles` calls.

The timeout for a SAML assertion is set to 5min as a constant, and the
code makes it relatively easy to make it configurable if needed, like
for example override this per provider.