Revert changes to time format, use named format specifier
Turns out the root cause was not what my changes addressed. My changes didn't break anything, but for the sake of clarity, I'm mostly reverting them. The only remaing changes
after this commit are to use a named format, rather than a confusing format string, and keeping the the correct order of arguments in the unit test.