Execute command directly without sh -c
When the args are concatenated there is a loss of fidelity:
arguments with spaces will be split and any special characters
will be evaluated by the shell.
If we simply execute the provided command it will propagate correctly
and it can always be wrapped in a shell on the command line
if that behavior is desired.

fixes #236