Fix nil dereference when JumpCloud DUO is denied or timed out
Response body will be closed in case of 401 by the time we attempt to
compare it with 200 in case if deny or time out. Not sure how it slipped
in, given I used existing code. Perhaps never worked or I made a silly
change somwhere, e.g. closing response body handler.