feat(keyring): ✨ Add support to set diff backend
Prior to this change, you were not able to set the keyring backend to
your preferred keyring. This mean it was up to `saml2aws` figure out what
the appropriate backend to use depending on the operating system, and a
priority list of backends that are supported.

After this change, you can use the environment variable
`SAML2AWS_KEYRING_BACKEND` to set it to whatever keyring backend you
want. However, this setting is only available for Linux operating
systems (unless this feature is required for other operating systems).

This hopefully allows headless systems such as WSL to set their backend
to a keyring that doesn't rely on X11 and DBus (such as pass).

This hopefully resolves the following issues:
- #563
- #582
- #561