feat: disambiguate Okta U2F tokens using the ID
This changes adds the ID of the token to the string in the prompt, which
allows users to distinguish tokens with non-unique authNames.

This also fixes the issue where if the authName is the same for all
tokens, the first token is always chosen by saml2aws regardless of user
selection.