Preserving previous behaviour for backward compatibility
This commit introduces a LegacyGet function for the keyring helpers.
This function provides the backward compatible behaviour we're looking
for while introducing a new keychain naming.

For most keyring providers, the function is the same, though called with
a different parameter name (idp name vs serverURL). However, for macs,
the behaviour is slightly different as it uses a common key name (so you
can never have more than one IdP).

Added some tests for the LookupCredentials function.