Okta + Duo: Allow passcodes from mobile app
Any Duo mobile app registered with Duo, usually for Duo Push, also
has the capability to produce HOTP passcodes. In the case where
Duo Push is administratively disabled and Yubikeys (etc) are not
provided, HOTP might be the only way for a user to perform Duo MFA.

This change adds Passcode to the list of MFA possibilities whenever a
phone1 is registered with Duo to support that situation.