fix: 🐛 simplify flow and remove password repeater (#1390)
Prior to this commit, the configure flow contained multiple if
statements, making the logic unnecessarily complex. Additionally, the
password repeater was included, which is not required as this process
is not a sign-up flow. The password repeater also caused ambiguity when
passwords did not match.

After this commit, The configure flow has been simplified by reducing
the number of if statements, making the logic more straightforward. The
password repeater has been removed entirely, as it is not needed. If the
user inputs the wrong password, they can simply re-initiate the
configure process.

fixes #1390

Signed-off-by: Sriram Venkatesh <sriram.venkatesh@versent.com.au>